requirements.txt
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
pip-chill==1.0.3
pyngrok==6.0.0
razorpay==1.3.0django==4.2.1
pip-chill==1.0.3
razorpay==1.3.0